Move exceptions to core library
