+++ ucp/_libs/ucx_worker_cb.pyx
@@ -10,9 +10,8 @@ from cython cimport boundscheck, initializedcheck, nonecheck, wraparound
 from libc.stdint cimport uintptr_t
 from libc.string cimport memcpy
 
+from .exceptions import UCXCanceled, UCXError, log_errors
 from .ucx_api_dep cimport *
-
-from ..exceptions import UCXCanceled, UCXError, log_errors
 
 logger = logging.getLogger("ucx")
